+++ tests/test_cfg_global.py
@@ -11,7 +11,7 @@ from logging import basicConfig
 from tempfile import TemporaryDirectory
 from timetracker.cfg.cfg_global import CfgGlobal
 from timetracker.cfg.cfg_local import CfgProj
-from timetracker.cfg.utils import get_relpath_adj
+#from timetracker.cfg.utils import get_relpath_adj
 from timetracker.cfg.utils import run_cmd
 from tests.pkgtttest.mkprojs import mkdirs
 from tests.pkgtttest.mkprojs import findhome
@@ -28,7 +28,7 @@ def test_cfgbase_home():
         assert cfg.doc['projects'] == []
         assert not exists(cfg.filename)
 
-def test_cfgbase_temp(name='taster', trksubdir='.timetracker'):
+def test_cfgbase_temp(trksubdir='.timetracker'):
     """Test cfg flow"""
     print(f'{SEP}1) INITIALIZE "HOME" DIRECTORY')
     with TemporaryDirectory() as tmphome:
@@ -45,17 +45,15 @@ def test_cfgbase_temp(name='taster', trksubdir='.timetracker'):
             # cfgname_proj = /tmp/tmptrz29mh6/proj/apples/.timetracker/config
             cfgname_proj = join(workdir, 'config')
             # EXP: apples '~/proj/apples/.timetracker/config'
-            exp_projs.append([proj, get_relpath_adj(cfgname_proj, tmphome)])
+            exp_projs.append([proj, cfgname_proj])
             # INIT LOCAL PROJECT CONFIG
-            cfgloc = CfgProj(cfgname_proj, project=proj, name=name)
+            cfgloc = CfgProj(cfgname_proj, project=proj)
             assert cfgloc.trksubdir == trksubdir, (f'\nEXP({trksubdir})\n'
                                                    f'ACT({cfgloc.trksubdir})\n'
                                                    f'{cfgloc}')
             assert cfgloc.dircfg == workdir
             assert cfgloc.project == proj
-            assert cfgloc.name == name
-            cfgloc.mk_dircfg()
-            cfgloc.wr_cfg_new()
+            cfgloc.write_file()
             # cat project/.timetracker/config
             filenamecfg_proj = cfgloc.get_filename_cfg()
             debug(f'PROJ CFG: {filenamecfg_proj}')