+++ timetracker/consts.py
@@ -6,6 +6,7 @@ __author__ = "DV Klopfenstein, PhD"
 
 DIRTRK = '.timetracker'
 DIRCSV = '.'
+FMTDT = '%Y-%m-%d %H:%M:%S.%f'
 
 
 # Copyright (C) 2025-present, DV Klopfenstein, PhD. All rights reserved.