+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Summary
 * [**Unreleased**](#unreleased)
+* **Release 2025-02-18 v0.2a4** Added `trk time` command to report elapsed hours
+* [**Release 2025-02-15 v0.2a3**](#release-2025-02-15-v02a3) Implemented functionality & Wrote tests for csv management
 * **Release 2025-02-08 v0.2a1** Cleaned up tests; Added pytest to GitHub workflows upon push
 * **Release 2025-02-07 v0.2a0** Added functionality for running in project subdirs
 * [**Release 2025-02-05 v0.1a8**](#release-2025-02-05-v01a8) Added DOI for citing
@@ -14,6 +16,13 @@
 # Details
 
 ## Unreleased
+
+## Release 2025-02-15 v0.2a3
+* ADD Test for csv management
+* ADD Finder function, get_dircsv_default, for printing in CLI help messages
+* ADD csv-finding functionality
+* CHANGE Starttime to its own module
+  * ADD CfgProj function to get Starttime instance
 
 ## Release 2025-02-05 v0.1a8
 * ADD: Zenodo citation and CITATION.cff file